+++ RELEASE_NOTES.md
@@ -38,6 +38,7 @@ Some things that are planned for the 7.0 release:
 ### ✨ New features
 
 * `KeyboardInputViewController` has a new `viewWillSyncWithContext` function.
+* `KeyboardTextContext` is a new observable class that lets you observe the document text.
 
 ### 👑 KeyboardKit Pro